Fix: lab6 - task1 - function problem c - switch to float

+++ Lab7/task1/informatics/5_func/b.py
@@ -1,5 +1,5 @@
 def power(a, n):
     return a ** n
 
-a, n = tuple(map(int, input().split()))
-print(power(a, n))
+a, n = tuple(map(float, input().split()))
+print(power(a, int(n)))
